Corregidos controladores de un objetivo estratégico para que se tengan en cuenta los ids de planes no existentes y se lance el respectivo error.

+++ icasus/app_code/objest_borrar.php
@@ -18,23 +18,32 @@ if (filter_has_var(INPUT_GET, 'id_objest') && filter_has_var(INPUT_GET, 'id_enti
     $id_entidad = filter_input(INPUT_GET, 'id_entidad', FILTER_SANITIZE_NUMBER_INT);
     $id_objest = filter_input(INPUT_GET, 'id_objest', FILTER_SANITIZE_NUMBER_INT);
 
+    //Obtenemos los datos del objetivo estratégico
     $objest = new ObjetivoEstrategico();
-    $objest->load("id=$id_objest");
-    $id_linea = $objest->id_linea;
-    $linea = new Linea();
-    $linea->load("id = $id_linea");
-    $objop = new ObjetivoOperacional();
-    $objop_objest = $objop->Find("id_objest=$id_objest");
-    if (!$objop_objest && $control)
+    if ($objest->load("id = $id_objest"))
     {
-        $exito = MSG_OBJEST_BORRADO . ' ' . $linea->indice . '.' . $objest->indice . '. ' . $objest->nombre;
-        $objest->delete();
-        header("Location: index.php?page=linea_mostrar&id_linea=$id_linea&id_entidad=$id_entidad&exito=$exito");
+        $id_linea = $objest->id_linea;
+        $linea = new Linea();
+        $linea->load("id = $id_linea");
+        $objop = new ObjetivoOperacional();
+        $objop_objest = $objop->Find("id_objest=$id_objest");
+        if (!$objop_objest && $control)
+        {
+            $exito = MSG_OBJEST_BORRADO . ' ' . $linea->indice . '.' . $objest->indice . '. ' . $objest->nombre;
+            $objest->delete();
+            header("Location: index.php?page=linea_mostrar&id_linea=$id_linea&id_entidad=$id_entidad&exito=$exito");
+        }
+        else
+        {
+            $aviso = MSG_OBJEST_BORRAR;
+            header("Location: index.php?page=objest_mostrar&id_objest=$id_objest&id_entidad=$id_entidad&aviso=$aviso");
+        }
     }
     else
     {
-        $aviso = MSG_OBJEST_BORRAR;
-        header("Location: index.php?page=objest_mostrar&id_objest=$id_objest&id_entidad=$id_entidad&aviso=$aviso");
+
+        $error = ERR_OBJEST_MOSTRAR;
+        header("location:index.php?page=error&error=$error");
     }
 }
 else
